requirements.txt
Flask==2.3.3
transformers==4.33.3
torch==2.0.1
tokenizers==0.13.3
python-telegram-bot==20.5